Solucionado que el billboard baje la vida poco a poco.
Mejora de la interfaz (nuevo estado InitState antes del Menu)

+++ Src/Logic/Entity/Components/Life.cpp
@@ -105,7 +105,7 @@ namespace Logic
 		}		
 
 		// LIFEBAR CONTROL
-		float ratio = _life / _LIFE_MAX;
+		float ratio = (float)_life / (float)_LIFE_MAX;
 			_lifeBar.setCoordenadas(
 				(1.0f - ratio) / 2.0f,			// u1
 				0.0f,							// v1
